refactor, drop app engine support.
Google is not betting on App Engine in the long run with dropped support
for memcache.
