--- deploy/app-engine/app.yaml
@@ -1,6 +1,0 @@
-runtime: go111
-
-handlers:
-- url: /
-  static_files: public/index.html
-  upload: public/(.*)
