--- deploy/app-engine/go.mod
@@ -1,8 +1,0 @@
-module github.com/jhaals/yopass/app-engine
-
-require (
-	github.com/gorilla/handlers v1.4.0
-	github.com/gorilla/mux v1.6.2
-	github.com/jhaals/yopass v0.0.0-20181017100916-1bd8e1a9928e
-	google.golang.org/appengine v1.2.0
-)
