+++ deploy/aws-lambda/go.mod
@@ -1,24 +1,16 @@
 module github.com/yopass/yopass-lambda
 
+replace github.com/jhaals/yopass => ../../
+
 require (
-	github.com/akrylysov/algnhsa v0.0.0-20190203201208-70f315bb89bd
-	github.com/aws/aws-lambda-go v1.8.2
-	github.com/aws/aws-sdk-go v1.16.26
+	github.com/akrylysov/algnhsa v0.12.1
+	github.com/aws/aws-lambda-go v1.13.2
+	github.com/aws/aws-sdk-go v1.25.8
+	github.com/bradfitz/gomemcache v0.0.0-20190913173617-a41fca850d0b // indirect
 	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/fnproject/fdk-go v0.0.0-20190102214815-bd24a5aa63cf
-	github.com/go-ini/ini v1.41.0 // indirect
-	github.com/gopherjs/gopherjs v0.0.0-20181103185306-d547d1d9531e // indirect
-	github.com/gorilla/context v1.1.1 // indirect
-	github.com/jhaals/yopass v0.0.0-20190204191800-daad8abb0cd4
-	github.com/jtolds/gls v4.2.1+incompatible // indirect
-	github.com/kr/pretty v0.1.0 // indirect
-	github.com/kr/pty v1.1.3 // indirect
-	github.com/smartystreets/assertions v0.0.0-20190116191733-b6c0e53d7304 // indirect
-	github.com/smartystreets/goconvey v0.0.0-20181108003508-044398e4856c // indirect
-	github.com/stretchr/objx v0.1.1 // indirect
-	github.com/stretchr/testify v1.3.0 // indirect
-	golang.org/x/net v0.0.0-20190125091013-d26f9f9a57f3 // indirect
-	golang.org/x/text v0.3.0 // indirect
-	gopkg.in/check.v1 v1.0.0-20180628173108-788fd7840127 // indirect
-	gopkg.in/ini.v1 v1.41.0 // indirect
+	github.com/jhaals/yopass v0.0.0-20191008072641-a6997e2f9a9a
+	golang.org/x/net v0.0.0-20191007182048-72f939374954 // indirect
 )
+
+go 1.13
